+++ class/helios/PESRetourCloudStorage.class.php
@@ -64,7 +64,7 @@ class PESRetourCloudStorage  implements ICloudStorable {
 	public function getFinder(): Finder
 	{
 		$finder = new Finder();
-		$finder->in($this->helios_responses_root)->name("*.xml");
+		$finder->in($this->helios_responses_root)->name("PES2R*.xml");
 		return $finder;
 	}