+++ class/helios/PesAllerStorage.class.php
@@ -10,18 +10,21 @@ class PesAllerStorage {
     private $heliosTransactionsSQL;
     private $openStackSwiftWrapper;
     private $logger;
+    private $repertoirePesAllerSansTransaction; // "/data/tdt-workspace/mail/helios_orphelins/"
 
 
     public function __construct(
         $helios_files_upload_root,
         HeliosTransactionsSQL $heliosTransactionsSQL,
         OpenStackSwiftWrapper $openStackSwiftWrapper,
-        Monolog\Logger $logger
+        Monolog\Logger $logger,
+        $repertoirePesAllerSansTransaction
     ) {
         $this->helios_files_upload_root = $helios_files_upload_root;
         $this->heliosTransactionsSQL = $heliosTransactionsSQL;
         $this->openStackSwiftWrapper = $openStackSwiftWrapper;
         $this->logger = $logger;
+        $this->repertoirePesAllerSansTransaction = $repertoirePesAllerSansTransaction;
     }
 
 	/**
@@ -111,7 +114,7 @@ class PesAllerStorage {
 	 * @param int $no_access_during_nb_days
 	 * @throws Exception
 	 */
-    public function menageLocal($no_access_during_nb_days = 9999, $do = true){
+    public function menageLocal($no_access_during_nb_days = 9999){
 
 		$sigtermHandler = SigTermHandler::getInstance();
         $dh = opendir($this->helios_files_upload_root);
@@ -130,17 +133,13 @@ class PesAllerStorage {
 				$this->logger->debug("File $file too young to die : not deleted");
 				continue;
             }
-            if (! $this->openStackSwiftWrapper->fileExistsOnCloud(
-                self::CONTAINER_NAME,
-                $file
-            )){
+            if (!$this->fileExistsOnCloud($file)){
             	$this->logger->info("File $file not existing on cloud : not deleted");
+                $this->handleOlderFileNotInCloud($file);
                 continue;
             }
 			$this->logger->info("Deleting file : $file");
-            if ($do) {
-				unlink($this->helios_files_upload_root . "/" . $file);
-			}
+			unlink($this->helios_files_upload_root . "/" . $file);
         }
         closedir($dh);
     }
@@ -151,6 +150,60 @@ class PesAllerStorage {
         $no_access_during_nb_seconds = $no_access_during_nb_days*86400;
         $this->logger->debug("Nombre de jour depuis la derniere modif : " . round($nb_seconds_without_access/60/60/24));
         return ($nb_seconds_without_access < $no_access_during_nb_seconds);
+    }
+
+    /**
+     * @param $file
+     */
+    private function moveToOrphelinsDirectory($file): void
+    {
+        if (is_null($this->repertoirePesAllerSansTransaction)) {
+            $this->logger->info(
+                "File $file : destination directory $this->repertoirePesAllerSansTransaction not found"
+            );
+            return;
+        }
+        if (!rename(
+            $this->helios_files_upload_root . "/" . $file,
+            $this->repertoirePesAllerSansTransaction . "/" . $file
+        )
+        ) {
+            $this->logger->info("File $file : rename KO");
+            return;
+        }
+    }
+
+    /**
+     * @param $file
+     */
+    private function handleOlderFileNotInCloud($file): void
+    {
+        $id = $this->heliosTransactionsSQL->getIdBySHA1($file);
+        if (!$id) {
+            $this->logger->info("$file No transaction id found");
+            $this->moveToOrphelinsDirectory($file);
+            return;
+        }
+        if (!$this->heliosTransactionsSQL->isTransactionAvailable($id)) {
+            $this->logger->info("$file [transaction $id] passé à not_available = false");
+            $this->heliosTransactionsSQL->setTransactionAvailable($id, true);
+        }
+        if ($this->heliosTransactionsSQL->isTransactionInCloud($id)) {
+            $this->logger->info("$file [transaction $id] passé à is_in_cloud = false");
+            $this->heliosTransactionsSQL->setTransactionInCloudRemove($id);
+        }
+    }
+
+    /**
+     * @param $file
+     * @return bool|\Psr\Http\Message\ResponseInterface
+     */
+    private function fileExistsOnCloud($file)
+    {
+        return $this->openStackSwiftWrapper->fileExistsOnCloud(
+            self::CONTAINER_NAME,
+            $file
+        );
     }
 
 }