+++ model/HeliosTransactionsSQL.class.php
@@ -375,6 +375,12 @@ class HeliosTransactionsSQL extends SQL {
         return $this->queryOne($sql,$sha1);
     }
 
+    public function isTransactionInCloud(int $object_id): bool
+    {
+        $sql = "SELECT is_in_cloud FROM helios_transactions WHERE id=?";
+        return  $this->queryOne($sql,$object_id);
+    }
+
     public function isTransactionAvailable(int $object_id): bool
     {
         $sql = "SELECT not_available FROM helios_transactions WHERE id=?";
@@ -404,6 +410,4 @@ class HeliosTransactionsSQL extends SQL {
         $sql = "SELECT id FROM helios_transactions WHERE acquit_filename=?";
         return $this->queryOne($sql,$pes_aquit_filename);
     }
-
-
 }