+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - passage du temps minimum d'éxécution des WorkerScript de 1s vers 10s
 
 ## Evolution
+- déplacement des PES Aller non liés à une transaction dans le répertoire HELIOS_PESALLER_SANSTRANSACTION
 - création d'un groupe supervisor mailsec-menage contenant les scripts cloud des mails #245
 - Intéger la valeur du codcol dans le mail listant les transactions Helios restées à l'état transmis #675